COC-192-chore(#135): 마이페이지 코드 임시 제거
