+++ src/pages/MyPage/MySpaceList/index.tsx
@@ -1,0 +1,5 @@
+import S from './style';
+
+export default function MySpaceList() {
+  return <S.Container>MySpaceList</S.Container>;
+}
