+++ src/pages/MyPage/MyStudyList/index.tsx
@@ -1,0 +1,5 @@
+import S from './style';
+
+export default function MyStudyList() {
+  return <S.Container>MyStudyList</S.Container>;
+}
