+++ src/pages/MyPage/MyStudyList/style.ts
